Move pkg/agent to plugins/pkg/agent
