+++ cmd/kubeapps-apis/plugins/pkg/agent/agent_test.go
