+++ cmd/kubeapps-apis/plugins/pkg/agent/docker_secrets_postrenderer_test.go
