+++ cmd/kubeapps-apis/plugins/pkg/agent/httpConfigForCluster.go
